Move from Travis to Jenkins (#178)
* Move from Travis to Jenkins

This moves the CI parts of the CI/CD pipeline from Travis to Jenkins. 

The Jenkins job currently runs on a dedicated machine and is not immutable, so it's theoretically possible for a previous build to break/influence the current one. 

It requires a working Xcode install (best to launch it once to "install additional components") and a swift-format binary (in /usr/local/bin) that matches the Xcode syntax parser thing.